backend/requirements.txt
Django==5.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
Markdown==3.5.2
python-decouple==3.8
python-dotenv==1.0.0

# HTTP i parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Production Server
gunicorn>=21.0.0

# PostgreSQL
psycopg2-binary>=2.9.9

# Static Files (prod)
whitenoise>=6.6.0

# Database URL parsing (Railway)
dj-database-url>=2.1.0
